+++ public/lancamento.php
@@ -52,29 +52,9 @@
                     <select class="input__box" name="categoria" id="categoria">
                         <option value="" default>Selecione</option>
                         <?php
-                        require '../src/get/categorias.php';
-
-                        function viewCategoria($categorias)
-                        {
-                            // Initialize an empty string for the <option> tags
-                            $options = '';
-
-                            // Loop through each item in the array
-                            foreach ($categorias as $item) {
-                                if (isset($item['id']) && isset($item['nome'])) {
-                                    // Append an <option> tag with the appropriate value and innerHTML
-                                    $options .= '<option value="' . htmlspecialchars($item['id']) . '">'
-                                        . htmlspecialchars($item['nome'])
-                                        . '</option>\n';
-                                }
-                            }
-
-                            // Displays option tags
-                            echo $options;
-                        }
                         try {
-                            $categorias = getCategorias();
-                            viewCategoria($categorias);
+                            include '../src/_view/categorias-as-options.php';
+                            getCategorias();
                         } catch (Throwable $err) {
                             echo $err->getMessage();
                         }
